Show correct recipe images in recipe cards
Showing dynamic images in React Native was kind of a pain to figure out, especially considering React Native works primarily with static images, but I'm glad I found a workaround that's still kind of robust. This solution mimics what happens with databases where rather than loading the recipe and the entire image JSON object, the image's id is used to get it from a separate collection/column depending on the type of database. Learned about using uri images too which was helpful